Fix: changes chunk_worker path to application root
The original relative path to load the chunkworker bundle doesn't work
in a single page application if the neuorglancer code isn't rendered at
the top level, because it will look for the chunk_worker in a path such
as /page/component/neuroglancer/chunk_worker.bundle.js, which the server
wont be able to find and can't return. Setting the path to
/chunk_worker.bundle.js means that it can be found by the server and
returned in a reliable fashion.