Fix: changes async_computation worker path
The original relative path to load the bundle doesn't work
in a single page application if the neuorglancer code isn't rendered at
the top level, because it will look for the bundle in a path such as
/page/component/neuroglancer/async_computation.bundle.js,
which the server wont be able to find and can't return. Setting
the path to start with '/' means that it can be found by the server and
returned in a reliable fashion.