Adds run time config for bundle file root paths.
Configuring the bundle loading path at build time was too restrictive as
it doesn't give the module user the ability to place the bundle files
where they want in their server tree. This change makes it possible for
them to specify the bundleRoot at runtime.The default behavior is to
serve the bundle files from the same relative location as the script.