Fix (ui/context_menu): Menu removed from page flow when hidden
Setting the visibility of the context menu div to hidden leaves it in
the page and affects the flow of content around it, which could have
undesired consequences. Setting the display to none removes it from the
page and stops rendering issues. With that in place, there is no need to
insert the context menu into a special container on the page, so the
code that checks for a target can be restored to a simple
document.body.appendChild() call.