Fix: context menu on right side of browser window
When the context menu is too close to the right side of the browser
window, it needs to know the offset, so that it can decide to move
it further into the page. This can't happen if the display is set to
none, because it is not rendered in the DOM and doesn't have an offset.