fix: disables right click context menu for module build.
The code that disables the context menu was not being called in the
module version. This meant that right clicking on the image to navigate
was interrupted by a pop menu from the browser. This update applies the
context menu disabling to the parent DOM node of the neuroglancer window
as opposed to using 'document'. This means we can still access the
context menu outside of neuroglancer in the parent application.